rename: lib 폴더 -> libs 네이밍 컨벤션 개선
