+++ airbnb/src/libs/Schedule/Components/Schedule.tsx
@@ -10,7 +10,6 @@ export function Schedule({
   endDate,
   setStartDate,
   setEndDate,
-  displaySchedule,
 }: ScheduleProps): JSX.Element {
   const { changeCurrentDate, currentDate } = useCurrentDate();
 
@@ -20,9 +19,8 @@ export function Schedule({
       endDate={endDate}
       setStartDate={setStartDate}
       setEndDate={setEndDate}
-      displaySchedule={displaySchedule}
     >
-      <S.ScheduleWrapper shouldDisplay={displaySchedule}>
+      <S.ScheduleWrapper>
         <S.BackButton type="button" onClick={changeCurrentDate(-1)}>
           <ArrowBackIosRoundedIcon fontSize="small" />
         </S.BackButton>
